Refactor Maps component to update type definitions and streamline data handling. Removed mock product data and integrated new structure for maps data, enhancing clarity and maintainability. Added backup file for previous implementation.
